requirements/base.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.3
h11==0.14.0
idna==3.8
mysql-connector-python==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
requests==2.32.3
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
